Refactor: Consolidate DOM utils and centralize UI icons
- Merged `content/core/dom-helpers.js` into `content/core/dom-utils.js` and updated all references.
- Created `content/components/ui/icons.js` as a shared library for SVG icons.
- Refactored `pages/blog/blog-app.js` and `pages/projekte/projekte-app.js` to use the shared icons.
- Updated `pages/projekte/projekte-app.js` to explicitly import React 19, resolving a version mismatch with global React 18.
- Updated `vite.config.js` manual chunks to reflect the file changes.
