+++ content/core/canonical-manager.js
@@ -1,5 +1,5 @@
 import { createLogger } from './logger.js';
-import { upsertHeadLink } from './dom-helpers.js';
+import { upsertHeadLink } from './dom-utils.js';
 
 const log = createLogger('CanonicalManager');
 const BASE_URL = 'https://www.abdulkerimsesli.de';
